improve: extract html transformation functions
